Add doc and snippets tests with Go runner (#244)
<!--
  This template is optional.
  It simply serves to provide a guide to allow a better review of pull requests.
-->

<!--
  IMPORTANT
  Don't forget to add the corresponding "changelog:xxx" label to your PR.
  This is part of our release process in order to generate the change log.
-->


## What does this PR do?
<!-- Please fill this section -->
This PR adds the doc in the `.doc` folder and snippet tests (with go runner)

### How should this be manually tested?

```
docker-compose -f .ci/doc/docker-compose.yml run doc-tests node index
```
### Other changes

1. I created a script names `formatAllSnippets.go` in order to solve an error thrown by the `go run` command. 
In each snippet's templates, we import several packages so that all the snippets that are using this template can work, though each of them are not using *all* of these imports. Thus, `go run` throws the following error : `package xxx was imported but not used`. 
To solve this, i used the command `goimports` that prints a formatted version of the snippet (free from all unused package). 
My script is iterating through each file, replacing it with the formatted version 

2. All occurrences of `for _ = range ch` have been replaced by `for range ch` since it caused an error during the lint phase (with `golint`)